refactor: fix clippy lints, remove spurious async from kafka stubs, and optimize barrier waker cleanup
Address accumulated clippy warnings across the service, messaging, and
sync layers:

- kafka.rs: remove async from send/flush/begin_transaction/commit/abort
  stubs that never await, fixing clippy::unused_async. Update all call
  sites (including tests) to drop .await on these calls.
- filter.rs: replace |_: &T| with |(): &()| to fix
  clippy::used_underscore_binding patterns; use assert!() instead of
  assert_eq!(x, true) for bool assertions.
- hedge.rs: use f64::EPSILON-based comparison instead of exact float eq;
  add post-clone assertions to prevent use-after-move regressions.
- steer.rs: fix wildcard binding pattern in test closure.
- barrier.rs: replace O(N) retain() fallback with position() +
  swap_remove() for waker cleanup on cancellation.
- lab/runtime.rs: remove redundant u128::from() cast.

Wrap database_pool_integration tests in a feature-gated mod tests block
(cfg sqlite/postgres/mysql) so they compile-skip cleanly when no DB
feature is active. Reformat grpc, quic, fs/signal, and I/O parity test
files for clippy and rustfmt compliance.

+++ src/lab/runtime.rs
@@ -1047,10 +1047,7 @@ impl LabRuntime {
         crate::tracing_compat::warn!(
             "virtual clock jump detected: old_time_ms={}, new_time_ms={}, jump_ms={} \
              -- may affect lease/timeout correctness",
-            self.virtual_time
-                .as_nanos()
-                .saturating_sub(u128::from(skew_nanos))
-                / 1_000_000,
+            self.virtual_time.as_nanos().saturating_sub(skew_nanos) / 1_000_000,
             self.virtual_time.as_nanos() / 1_000_000,
             skew_nanos / 1_000_000
         );
